+++ CHANGELOG.md
@@ -6,4 +6,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.1.0] - unreleased
 ### Added
-se/INCORE1-1273)
+- se/INCORE1-1273)
+- Data preprocessing steps till step 4 in guide
+